feat(io): Implement HDF5 property output framework
This commit is a major refactor of the property output system. A preliminary implementation of the HDF5 format is available.

Major changes:

Refactored IO classes

LocalPropertyOutput has been refactored to an abstracted base class

The Xdr writing logic has been moved to a dedicated XdrPropertyOutput subclass

A new LocalPropertyHdf5Output subclass has been created to handle hdf5 writing

XML config:

It now accepts a  attribute in the  tag

Such as format=hdf5 or format=xtr

SimConfig has been updated to parse this new property

PropertyWriter works as Factory Pattern

CMakeList has not been updated yet.
